chore(scorchwrap): refactor assemble functionality to a separate file
